Changed rtc::getTemperature to an in16_t (-214 bytes)
With this change, EEPROM historic data is unusable so version bump + config seed change

+++ src/Rtc.h
@@ -5,7 +5,7 @@
 namespace rtc {
 	void setup();
 
-	float getTemperature();
+	int16_t getTemperature();
 
 	bool isAccurate();
 	timestamp_t getTime();
